Move mocked client generation for import in all tests
